% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_mse}
\alias{compute_mse}
\title{Compute Mean Squared Error}
\usage{
compute_mse(reference, estimate)
}
\arguments{
\item{reference}{Reference SeismicDataset}

\item{estimate}{Estimated SeismicDataset}
}
\value{
MSE value
}
\description{
MSE = mean((reference - estimate)^2)
}
