% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_relative_error}
\alias{compute_relative_error}
\title{Compute Relative Error}
\usage{
compute_relative_error(reference, estimate)
}
\arguments{
\item{reference}{Reference SeismicDataset}

\item{estimate}{Estimated SeismicDataset}
}
\value{
Relative Frobenius norm error
}
\description{
RelError = ||estimate - reference||_F / ||reference||_F
}
