% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_snr}
\alias{compute_snr}
\title{Compute Signal-to-Noise Ratio}
\usage{
compute_snr(reference, estimate)
}
\arguments{
\item{reference}{Reference SeismicDataset (ground truth)}

\item{estimate}{Estimated/reconstructed SeismicDataset}
}
\value{
SNR in decibels
}
\description{
SNR = 10 * log10(P_signal / P_noise)
}
\examples{
ref <- SeismicDataset(matrix(rnorm(100), 10, 10), dt = 0.004)
est <- SeismicDataset(ref$traces + 0.1 * rnorm(100), dt = 0.004)
snr <- compute_snr(ref, est)

}
