% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{interpolate_at}
\alias{interpolate_at}
\title{Bilinear interpolation of velocity at position}
\usage{
interpolate_at(vm, x, z)
}
\arguments{
\item{vm}{VelocityModel object}

\item{x}{Horizontal position (m)}

\item{z}{Vertical position (m)}
}
\value{
Interpolated velocity value
}
\description{
Bilinear interpolation of velocity at position
}
