% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{promethium_load_hdf5}
\alias{promethium_load_hdf5}
\title{Load HDF5 File}
\usage{
promethium_load_hdf5(path, group = "seismic")
}
\arguments{
\item{path}{Path to HDF5 file}

\item{group}{Group name within HDF5 file}
}
\value{
SeismicDataset object
}
\description{
Read seismic dataset from HDF5 file format.
Requires hdf5r package.
}
