% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{promethium_load_segy}
\alias{promethium_load_segy}
\title{Load SEG-Y File}
\usage{
promethium_load_segy(path)
}
\arguments{
\item{path}{Path to SEG-Y file}
}
\value{
SeismicDataset object
}
\description{
Read seismic data from SEG-Y file format.
Note: This is a simplified implementation supporting common variants.
}
