% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{promethium_run}
\alias{promethium_run}
\title{Run Recovery Pipeline}
\usage{
promethium_run(pipeline, data, mask = NULL, verbose = FALSE)
}
\arguments{
\item{pipeline}{RecoveryPipeline object}

\item{data}{SeismicDataset object}

\item{mask}{Logical matrix (TRUE = observed, FALSE = missing)}

\item{verbose}{Print progress}
}
\value{
Recovered SeismicDataset
}
\description{
Execute pipeline on seismic data with optional mask for missing data.
}
\examples{
ds <- promethium_synthetic(ntraces = 20, nsamples = 100)
pipeline <- from_preset("wiener")
result <- promethium_run(pipeline, ds)

}
