% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{remove_dc}
\alias{remove_dc}
\alias{remove_dc.numeric}
\alias{remove_dc.SeismicDataset}
\title{Remove DC Offset}
\usage{
remove_dc(x)

\method{remove_dc}{numeric}(x)

\method{remove_dc}{SeismicDataset}(x)
}
\arguments{
\item{x}{Numeric vector or SeismicDataset}
}
\value{
Signal with zero mean
}
\description{
Remove mean (DC component) from signal.
}
